correções

+++ src/main/java/io/github/brunaoflofl/clientes/model/repository/ServicoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.brunaoflofl.clientes.model.entity.Servico;
 
-public interface ServicoRepository extends JpaRepository<Servico, Integer> {
+public interface ServicoRepository extends JpaRepository<Servico, Long> {
 
 }
